+++ pkg/statistics/handle/BUILD.bazel
@@ -5,7 +5,6 @@ go_library(
     srcs = [
         "bootstrap.go",
         "ddl.go",
-        "dump.go",
         "handle.go",
         "handle_hist.go",
     ],
@@ -37,7 +36,6 @@ go_library(
         "//pkg/types",
         "//pkg/util",
         "//pkg/util/chunk",
-        "//pkg/util/intest",
         "//pkg/util/logutil",
         "//pkg/util/sqlexec",
         "@com_github_pingcap_errors//:errors",
@@ -53,7 +51,6 @@ go_test(
     timeout = "short",
     srcs = [
         "ddl_test.go",
-        "dump_test.go",
         "gc_test.go",
         "handle_hist_test.go",
         "main_test.go",
@@ -61,23 +58,17 @@ go_test(
     embed = [":handle"],
     flaky = True,
     race = "on",
-    shard_count = 27,
+    shard_count = 13,
     deps = [
         "//pkg/config",
-        "//pkg/domain",
         "//pkg/parser/model",
         "//pkg/planner/cardinality",
         "//pkg/sessionctx",
         "//pkg/sessionctx/stmtctx",
         "//pkg/sessionctx/variable",
-        "//pkg/statistics",
-        "//pkg/statistics/handle/internal",
-        "//pkg/statistics/handle/storage",
-        "//pkg/statistics/handle/util",
         "//pkg/testkit",
         "//pkg/testkit/testsetup",
         "//pkg/types",
-        "//pkg/util",
         "//pkg/util/mathutil",
         "//pkg/util/mock",
         "@com_github_pingcap_failpoint//:failpoint",
