+++ pkg/statistics/handle/storage/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "//pkg/types",
         "//pkg/util/chunk",
         "//pkg/util/compress",
+        "//pkg/util/intest",
         "//pkg/util/logutil",
         "//pkg/util/mathutil",
         "//pkg/util/memory",
@@ -45,13 +46,23 @@ go_library(
 go_test(
     name = "storage_test",
     timeout = "short",
-    srcs = ["read_test.go"],
+    srcs = [
+        "dump_test.go",
+        "read_test.go",
+    ],
     flaky = True,
+    shard_count = 16,
     deps = [
+        ":storage",
+        "//pkg/domain",
         "//pkg/parser/model",
         "//pkg/planner/cardinality",
+        "//pkg/statistics",
+        "//pkg/statistics/handle/internal",
+        "//pkg/statistics/handle/util",
         "//pkg/testkit",
         "//pkg/types",
+        "//pkg/util",
         "@com_github_stretchr_testify//require",
     ],
 )
