+++ pkg/statistics/handle/storage/stats_read_writer.go
@@ -15,16 +15,22 @@
 package storage
 
 import (
+	"context"
 	"fmt"
+	"runtime"
+	"sync"
+	"sync/atomic"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/pkg/config"
+	"github.com/pingcap/tidb/pkg/infoschema"
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
 	"github.com/pingcap/tidb/pkg/statistics"
 	handle_metrics "github.com/pingcap/tidb/pkg/statistics/handle/metrics"
 	"github.com/pingcap/tidb/pkg/statistics/handle/util"
+	"github.com/pingcap/tidb/pkg/util/intest"
 	"github.com/pingcap/tidb/pkg/util/sqlexec"
 )
 
@@ -357,4 +363,127 @@ func (s *statsReadWriter) TableStatsToJSON(dbName string, tableInfo *model.Table
 		return nil, err
 	}
 	return jsonTbl, nil
+}
+
+// TestLoadStatsErr is only for test.
+type TestLoadStatsErr struct{}
+
+// LoadStatsFromJSON will load statistic from JSONTable, and save it to the storage.
+// In final, it will also udpate the stats cache.
+func (s *statsReadWriter) LoadStatsFromJSON(ctx context.Context, is infoschema.InfoSchema,
+	jsonTbl *util.JSONTable, concurrencyForPartition uint8) error {
+	if err := s.LoadStatsFromJSONNoUpdate(ctx, is, jsonTbl, concurrencyForPartition); err != nil {
+		return errors.Trace(err)
+	}
+	return errors.Trace(s.statsHandler.Update(is))
+}
+
+// LoadStatsFromJSONNoUpdate will load statistic from JSONTable, and save it to the storage.
+func (s *statsReadWriter) LoadStatsFromJSONNoUpdate(ctx context.Context, is infoschema.InfoSchema,
+	jsonTbl *util.JSONTable, concurrencyForPartition uint8) error {
+	nCPU := uint8(runtime.GOMAXPROCS(0))
+	if concurrencyForPartition == 0 {
+		concurrencyForPartition = nCPU / 2 // default
+	}
+	if concurrencyForPartition > nCPU {
+		concurrencyForPartition = nCPU // for safety
+	}
+
+	table, err := is.TableByName(model.NewCIStr(jsonTbl.DatabaseName), model.NewCIStr(jsonTbl.TableName))
+	if err != nil {
+		return errors.Trace(err)
+	}
+	tableInfo := table.Meta()
+	pi := tableInfo.GetPartitionInfo()
+	if pi == nil || jsonTbl.Partitions == nil {
+		err := s.loadStatsFromJSON(tableInfo, tableInfo.ID, jsonTbl)
+		if err != nil {
+			return errors.Trace(err)
+		}
+	} else {
+		// load partition statistics concurrently
+		taskCh := make(chan model.PartitionDefinition, len(pi.Definitions))
+		for _, def := range pi.Definitions {
+			taskCh <- def
+		}
+		close(taskCh)
+		var wg sync.WaitGroup
+		e := new(atomic.Pointer[error])
+		for i := 0; i < int(concurrencyForPartition); i++ {
+			wg.Add(1)
+			s.statsHandler.GPool().Go(func() {
+				defer func() {
+					if r := recover(); r != nil {
+						err := fmt.Errorf("%v", r)
+						e.CompareAndSwap(nil, &err)
+					}
+					wg.Done()
+				}()
+
+				for def := range taskCh {
+					tbl := jsonTbl.Partitions[def.Name.L]
+					if tbl == nil {
+						continue
+					}
+
+					loadFunc := s.loadStatsFromJSON
+					if intest.InTest && ctx.Value(TestLoadStatsErr{}) != nil {
+						loadFunc = ctx.Value(TestLoadStatsErr{}).(func(*model.TableInfo, int64, *util.JSONTable) error)
+					}
+
+					err := loadFunc(tableInfo, def.ID, tbl)
+					if err != nil {
+						e.CompareAndSwap(nil, &err)
+						return
+					}
+					if e.Load() != nil {
+						return
+					}
+				}
+			})
+		}
+		wg.Wait()
+		if e.Load() != nil {
+			return *e.Load()
+		}
+
+		// load global-stats if existed
+		if globalStats, ok := jsonTbl.Partitions[util.TiDBGlobalStats]; ok {
+			if err := s.loadStatsFromJSON(tableInfo, tableInfo.ID, globalStats); err != nil {
+				return errors.Trace(err)
+			}
+		}
+	}
+	return nil
+}
+
+func (s *statsReadWriter) loadStatsFromJSON(tableInfo *model.TableInfo, physicalID int64, jsonTbl *util.JSONTable) error {
+	tbl, err := TableStatsFromJSON(tableInfo, physicalID, jsonTbl)
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	for _, col := range tbl.Columns {
+		// loadStatsFromJSON doesn't support partition table now.
+		// The table level count and modify_count would be overridden by the SaveMetaToStorage below, so we don't need
+		// to care about them here.
+		err = s.SaveStatsToStorage(tbl.PhysicalID, tbl.RealtimeCount, 0, 0, &col.Histogram, col.CMSketch, col.TopN, int(col.GetStatsVer()), 1, false, util.StatsMetaHistorySourceLoadStats)
+		if err != nil {
+			return errors.Trace(err)
+		}
+	}
+	for _, idx := range tbl.Indices {
+		// loadStatsFromJSON doesn't support partition table now.
+		// The table level count and modify_count would be overridden by the SaveMetaToStorage below, so we don't need
+		// to care about them here.
+		err = s.SaveStatsToStorage(tbl.PhysicalID, tbl.RealtimeCount, 0, 1, &idx.Histogram, idx.CMSketch, idx.TopN, int(idx.GetStatsVer()), 1, false, util.StatsMetaHistorySourceLoadStats)
+		if err != nil {
+			return errors.Trace(err)
+		}
+	}
+	err = s.SaveExtendedStatsToStorage(tbl.PhysicalID, tbl.ExtendedStats, true)
+	if err != nil {
+		return errors.Trace(err)
+	}
+	return s.SaveMetaToStorage(tbl.PhysicalID, tbl.RealtimeCount, tbl.ModifyCount, util.StatsMetaHistorySourceLoadStats)
 }
